--- src/main/java/frc/robot/subsystems/LimeVision/Testing.txt
@@ -1,12 +1,0 @@
-GET <http://<IPADDRESS>:5807/results> 	
-Retreive full JSON dump of current targeting results
-
-GET <http://<IPADDRESS>:5807/capturesnapshot> 	
-Capture a snapshot. Include a “snapname” string header to name the snapshot
-
-POST <http://<IPADDRESS>:5807/uploadsnapshot> 	
-Upload a named snapshot. Include a “snapname” header. Include an image in the body of the request
-
-
-ws://<IPADDRESS>:5806 	
-Websocket server streaming full JSON dump of current targeting results at full framerate
